refactor: explicit configureStore for dev/prod env
